clean design for createvariant usecase

+++ database/migrations/2023_09_14_104129_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('description')->nullable();
             $table->timestamps();
-            $table->decimal('default_price', 9,2)->nullable();
+            $table->integer('default_price')->nullable();
             $table->integer('default_weight')->nullable();
         });
     }
