+++ database/migrations/2023_09_18_162628_create_variants_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
                 ->on('products')->onDelete('cascade');
             $table->string('name');
             $table->string('sku');
-            $table->decimal('price', 9,2)->nullable();
+            $table->integer('price')->nullable();
             $table->integer('weight')->nullable();
         });
     }
